Add files via upload
Photo 1 shows the basic look of the game after clicking the upper button on the first form that appears: Roman units are red and Parthian units are purple.
Photo 2 shows how the basic interface of an engagement looks like. You can see how much manpower and morale was lost on the side of the attacker (left-hand side) and how much was lost on the side of the defender in an engagement (right-hand side).